Revert "Add Auth0"
