+++ docs/NOTES.md
@@ -8,8 +8,6 @@
 
 ### Features
 
-- AUTHENTICATION
-
 - Hide text box for click to invite a friend button ✅
 
 - Eliminate exchange letters bug by adding Stages to gameplay
