+++ package.json
@@ -7,7 +7,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/auth0-js": "^9.14.2",
     "@types/bluebird": "^3.5.32",
     "@types/jest": "^24.9.1",
     "@types/koa": "^2.11.3",
@@ -22,7 +21,6 @@
     "@types/sequelize": "^4.28.9",
     "@types/sqlite3": "^3.1.6",
     "@types/validator": "^13.0.0",
-    "auth0-js": "^9.14.3",
     "axios": "^0.19.2",
     "boardgame.io": "0.42.2",
     "dotenv": "^8.2.0",
@@ -39,7 +37,6 @@
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "4.0.1",
-    "react-use-auth": "^2.1.0-14",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^5.21.11",
     "sequelize-typescript": "^1.1.0",
