+++ src/webapp/App.test.tsx
@@ -4,12 +4,12 @@ import { Provider } from "react-redux";
 import { store } from "./app/store";
 import App from "./App";
 
-test.skip("broken due to rabble logo updates, TODO", () => {
-
-  const { getAllByText } = render(
+test("renders learn react link", () => {
+  const { getByText } = render(
     <Provider store={store}>
       <App />
     </Provider>
   );
-  expect(getAllByText(/Rabble/i).length).toBeGreaterThan(0);
+
+  expect(getByText(/Rabble/i)).toBeInTheDocument();
 });
