--- src/webapp/features/auth/auth.module.css
@@ -1,87 +1,0 @@
-.container {
-  background-color: #fffeff;
-  text-align: center;
-  width: 100%;
-  height: 75vh;
-  padding: 1rem 3.5rem;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.853);
-}
-
-.container > h1 {
-  font-size: 50px;
-  font-family: var(--font-swash);
-  margin-bottom: 15%;
-}
-
-.container > p {
-  font-style: italic;
-  color: rgba(0, 0, 0, 0.637);
-}
-
-.container > button {
-  width: 100%;
-  height: 55px;
-  color: white;
-  border-radius: 5px;
-  background-color: var(--btn-base);
-}
-.container > button > span {
-  font-size: 20px;
-}
-
-.loginTilesContainer {
-  margin: 10% auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 50px;
-  position: relative;
-}
-
-.loginTile {
-  font-size: 28px;
-  font-style: bold;
-  font-family: "Lato", sans-serif;
-  display: inherit;
-  justify-content: inherit;
-  align-items: inherit;
-  width: 50px;
-  height: 50px;
-  border-radius: 5px;
-  background-image: url("../../features/rabble/components/blank.png");
-  background-position: center;
-  background-size: cover;
-  box-shadow: 1px 1px 2px black;
-}
-.loginTile:nth-of-type(1) {
-  transform: rotateZ(-3deg);
-}
-.loginTile:nth-of-type(2) {
-  transform: rotateZ(4deg);
-}
-
-@media only screen and (min-width: 600px) {
-  .container {
-    width: 600px;
-    height: 700px;
-    margin: 10vh auto;
-    padding: 4rem 3.5rem;
-    border-radius: 15px;
-    box-shadow: 5px 7px 15px rgb(120, 110, 110);
-  }
-
-  .container > p,
-  .container > button {
-    margin-top: 10%;
-  }
-
-  .container > h1 {
-    padding: 7% 5%;
-    border-radius: 5px;
-    margin: 0 auto;
-  }
-  .loginTilesContainer {
-    margin: 8% auto 3% auto;
-  }
-}
