+++ src/webapp/features/rabble_logo/rabbleLogo.tsx
@@ -3,12 +3,8 @@ import { useEffect, useRef } from "react";
 import styles from "./rabbleLogo.module.css";
 import LogoCube from "./logoCube";
 
-import { useAuth } from "react-use-auth";
-
 const RabbleLogo = () => {
   const containerRef = useRef<HTMLInputElement>(null);
-
-  const { isAuthenticated } = useAuth();
 
   // the following useEffect and onSkipBtn disp:none the animation container
   // to reduce negative interaction likelihood.
@@ -24,16 +20,12 @@ const RabbleLogo = () => {
     };
   }, []);
 
-  if (isAuthenticated()) {
-    return null;
-  }
-
   const onSkipBtn = () => {
     const skipStyle = containerRef?.current?.style;
     if (skipStyle) {
       skipStyle.display = "none";
     }
-  }
+  } 
 
   return (
     <>
